Extract native file parser to machinefile source file.
